Excise now-used RootState for "genomics" summary views, and rename containing file accordingly
