+++ Site/webapp/wdkCustomization/js/client/controllers/GenomeSummaryViewController.tsx
@@ -25,9 +25,15 @@ import {
 
 import { GenomeSummaryView } from '../components/genomeSummaryView/GenomeSummaryView';
 
-import { RootState, GenomeSummaryViewReport } from '../types/summaryViewTypes';
+import * as genomeSummaryViewStoreModule from '../storeModules/GenomeSummaryViewStoreModule';
+
+import { GenomeSummaryViewReport } from '../types/genomeSummaryViewTypes';
 
 import { GenomeSummaryViewReportModel, toReportModel } from '../util/GenomeSummaryViewUtils';
+
+interface StateSlice {
+  [genomeSummaryViewStoreModule.key]: genomeSummaryViewStoreModule.State
+}
 
 type StateProps = 
   | { status: 'loading' }
@@ -112,9 +118,8 @@ const reportModel = createSelector<GenomeSummaryViewReport, GenomeSummaryViewRep
   toReportModel
 );
 
-function mapStateToProps(state: RootState, props: OwnProps): StateProps {
+function mapStateToProps(state: StateSlice, props: OwnProps): StateProps {
   const genomeSummaryViewState = state.genomeSummaryView[props.viewId];
-  const globalDataState = state.globalData;
 
   if (genomeSummaryViewState == null) return { status: 'loading' };
 
@@ -146,7 +151,7 @@ function mapDispatchToProps(dispatch: Dispatch, props: OwnProps): DispatchProps 
   }, dispatch);
 }
 
-export default connect<StateProps, DispatchProps, OwnProps, Props, RootState>(
+export default connect<StateProps, DispatchProps, OwnProps, Props, StateSlice>(
   mapStateToProps,
   mapDispatchToProps,
   (state, actionCreators, ownProps) => ({ state, actionCreators, ownProps })
